bug fix
git-svn-id: http://istr.iis.u-tokyo.ac.jp/svn/istr/trunk@168 575e276f-71f8-4965-9d1f-d4eeffdffdfb

+++ fistr1/tools/neu2fstr/NFD/CNFDataBlock.h
@@ -31,7 +31,7 @@
 #endif
 
 
-typedef long int      nf_int;   // 4 byte
+typedef int           nf_int;   // 4 byte
 typedef unsigned char nf_bool;  // 1 byte
 typedef double        nf_float; // 8 byte
 typedef char          nf_char;  // 1 byte
